Simplification of server fetching by range so the store doesn't have excess work
